requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
astroid==2.15.8
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
cachetools==5.3.3
cffi==1.16.0
cleanlab==2.6.3
continuum==1.2.7
contourpy==1.2.0
cycler==0.12.1
datasets==2.18.0
decorator==5.1.1
dill==0.3.8
docopt==0.6.2
dodgy==0.2.1
exceptiongroup==1.2.0
fairscale==0.4.13
fire==0.6.0
flake8==5.0.4
flake8-polyfill==1.0.2
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
gitdb==4.0.11
gitpython==3.1.42
google-auth==2.29.0
google-auth-oauthlib==0.4.6
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.22.1
hydra-core==1.2.0
imagehash==4.3.1
imageio==2.34.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
iniconfig==2.0.0
isort==5.13.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
lazy-loader==0.4
lazy-object-proxy==1.10.0
librosa==0.10.1
llvmlite==0.42.0
markdown==3.6
matplotlib==3.8.3
mccabe==0.7.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy==1.9.0
mypy-extensions==1.0.0
numba==0.59.1
numpy==1.24.4
oauthlib==3.2.1
omegaconf==2.2.3
open-clip-torch==1.3.0
packaging==24.0
pandas==1.4.3
pep8-naming==0.10.0
pillow==9.2.0
pipreqs==0.4.11
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
prospector==1.10.3
protobuf==3.19.6
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycodestyle==2.9.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==2.5.0
pylint==2.17.7
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyparsing==3.1.2
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.0.post2
pytz==2024.1
pywavelets==1.5.0
regex==2023.12.25
requests-oauthlib==2.0.0
requirements-detector==1.2.2
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
semver==3.0.2
sentencepiece==0.2.0
setoptconf-tmp==0.3.1
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
soundfile==0.12.1
soxr==0.3.7
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.4.9
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
tornado==6.4
tqdm==4.64.0
transformers==4.21.1
visdom==0.2.4
wcwidth==0.2.13
websocket-client==1.8.0
werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
zipp==3.18.1